requirements.txt
tqdm
numpy
scipy
scikit-learn
faiss-cpu